Added waitForRequest built-in tasks
